study-loan-backend/requirements.txt
alembic==1.14.1
blinker==1.8.2
boto3==1.37.38
botocore==1.37.38
certifi==2026.1.4
cffi==1.17.1
click==8.1.8
cloudinary==1.44.1
cryptography==46.0.3
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==2.1.5
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
s3transfer==0.11.5
sendgrid==6.12.5
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.13.2
urllib3==1.26.20
Werkzeug==3.0.6
zipp==3.20.2
